Support to query connected and id vendor and product
Still a WIP, but trying to add support for knowing when a device is connected and when it disconnects.

Put in two places.  Main USB device class, plus HID input class.

Test program shows when connect and disconnect. Appears to work for Mouse and at times with Joystick.

There is an issue  with some disconnects not fully working.  Appears to be problem for Keyboards,, so also problem for combo units.