Use specific driver - before generic HID driver
Ran into issue, where  you plug in keyboard and it uses the keyboard driver as that was the first one installed.  You unplug and the keyboard driver is put to bottom of list.  You connect it or another keyboard (currently it will use the 2nd keyboard driver we had in test app), but on third one, it would grab HID driver and not work...

This change is to make the pass through interfaces in two loops (pass 1 and then 2 to claim method.  This allows the  keyboard driver and maybe other specific ones like if we do PS3 specific one to claim it first and if these fail, take second pass where HID driver will try to claim it

USBHidParser as keyword.

Hid driver now tries to claim objects on 2nd pass, allowing keyboard to
grab an appropriate object even if it is later in the list.