Fix enumeration issue.  When the device added has multiple interfaces with different drivers.  The code was only properly linking in the last driver.  So fixed the linking in.  This helped with both the Logitech and Microsoft Wireless Mouse and Keyboard.
Experimental: Read Device strings on demand

This is a WIP, currently doing this Async and using Event Responder.  To do this the HID Input class now holds onto pointer to base object so it can ask for the data.

The code is also doing a pretty limited attempt at converting the
returned string objects which start with a length, followed by 0x3,
followed by unicode strings, to an a zero teriminated Ascii string.

With the unicode characters it is simply ignoring the high order byte of
each character.

The code is semi limited as it only allows each object to have one query
active at a time, as I am using the enumeration code (state 42), to
handle when the packet comes back from the USB query.