This is the start of some keyboard updates
Added member to the keyboard class to allow the user to control the LEDS
on the keyboard.

So far it works with the test program probably needs some additional
work, like detecting when it is called from a callback function.  We may
need to detect this and postpone the change until the callback has
completed.

The Ctrl keys (both left and right)  is supposed to zero the top 2 bits (of 7 bit code) i.e. & 0x1f so added that.

Also to debug stuff have the on key print function also output the modifiers and oem key