Work around for MS keyboard
I updated the keyboard extras to detect if the report ID is 0xff00 and if so ignore it in the Mouse extras process code.  This helped not handling a lot of extra messages generated by MS keyboard.

Also updated Test app to show names of some of these special keys.