Fix delete delete hang on serial ports
Found the issue where it was freeing transfers from the follow up list, that were also on the QH list which also tried to free it again.

Fix was in the first free list, it checks to see if the item is also in the qh list and does not free it then.