PS3 Support - like Pair, plus try to get remote name
Support to pair PS3 if it knows the BTADDR and user presses select and PS button and plugged in USB.

Remembers remote name, and tries to get it when PAIR version of code is running.  (Currently fails).

Does some mapping of PS3 Bluetooth data to match USB HID data.

Warning: as part of debuggigng why joystickType value was changing, I changed the class to make it a function you call instead of just a member variable.  Should mark it inline...  I think I caught all of the places in example code that was accessing it.  If we find we want ability for user probram to update it, should add another member to explicityly set it.