WIP for Nitindo Switch controller USB ...
WIP -  Not complete, but beginning to have it start to work when plugged in USB.

Needed to skip some descriptors to find the end point descriptor.  Maybe should clean up some as to remove other special case.

Next need to understand the report data as it differs for normal XBOX360...  Note It says it is an Xbox 360...   Now it shows debug data with the data.  Need to decide how best to know the data and inform the ;host code of the new messages