PS4 BT - Map buttons same as USB - Switch controller - WIP sound/Leds...
Rework the mapping of PS4 buttons field in Bluetooth mode, to be the same as in USB mode.  ALthough I left in code to map the HAT buttons to upper word of Butons.  Maybe sometime might do same for USB/HID connection of PS4.

WIP for Swtitch controller to try to get leds and rumble to do something.