USB HID Info Example sketch
here have been many cases where users plug in some new HID device, which we do not currently support, and we have very little information to go on.

So this sketch simply tries to print out the HID information for each message received from the device.

It has a simple sub-class of the USBHIDInput class, which prints out the data.  Which could give us a lot of data on adding support to the library, plus shows how users can write their own sub-class for their own devices.

This also includes a class to display information about the
device that was plugged in.  A lot of this is similar to
the data you would see if you turned debug on to the library

I thought maybe showing additional information about the device, like the debug output version of library might help get more detailed information back from users who have devices that are not supported/working