USBSerial fix add USBSerialEmu
I was workign on some other USBHost extensions and found I needed debug information that was
being generated by a Teensy plugged into the USB host port that was output using Serial Emulation
to be available.  Likewise I wanted the ability to feed data back to that other Teensy.

So I decided to create USBHost_t36 implemention to handle the USB types that include Serial Emulation
USBSerialEmu class.

The USBSerialEmu code was based on the USBSerial code as well as RAWHID.  I found that there was a bug
in the Receiving of data that was sent by the remote SEREMU output, I would miss the first character
Fixed.

USBSerial fixed same bug in reading first character