Experiment - Add a msReadSectorsWithCB function
Experiment with the VolumeName and other MSC example scketchs in the
UsbMSCFat project, in order to get the free space faster!

This included adding timeouts as sometimes the host does not fully respond.
Try to keep the code from hanging if for some reason we don't get enough messages back from host...