Bluetooth PS4 sketch and supporting some clones
With previous verison was able to get the Sony PS4 to work
but not most of the clones including one by Voyee.

Found that if I disabled some code that was put in for PS4
to force our code to initiate the bind operation.  Then the
voyee worked... Found later that in the USB Host Shield 2
that they detected the underlying class number and only if
0x2508 did it initiate the bind... voyee is 0x508.  So both
now appear to want to bind.

Changed the JoystickBT sketch to allow you to hit enter in debug
window and switch PS4 from long view to short view of data

Also found information on PS4 binding and some Python code that
allowed PS4s to bind without having to do it through BT.  So added
code like we have for PS3 that allows you to first have
the BT dongle in, then plus in PS4, and then hold either L1 or R1 and
hit the Share button and it will try to update the PS4 with the pairing
information for your BT dongle.

During this found I was displaying the BDADDR byte wise backwards so
updated this sketch.