PFSLib - F(msgs), fix m_pr usages, detect GPT in mbrdmp
Started off wanting to detect that it is a GPT drive instead of MBR drive when I do a dump of partitions.  Will want that elsewhere as well..

Decide to move most of the messages in the file to F("...") strings as on T4.x don't need to have strings in DTCM...

Also found some places where not using the stream passed in to print.