Start enumeration support plus
First pass to add the functionality we were using with the previous
integration pass and work toward beta 2...

Still retaining some support for the Beta1 way of starting up a filesystem:

but added USBFilesystem::end() which closes the file system allows you to reuse the object.

Added bool operator to USBFilesystem know if object is valid.

Added USBFSBase class that works like hid class or the like, that USBFilesystem derives from. Which allows us to have
multiple Filesystem implementations.

Added code to create list of these, with ini() method added to Filesystem object to add it to a free list.
When these objects are claimed, I remove them from this list and add them to list connected to drive.  When drive
goes away, I end() these objects and then put them back on free list.  Note: I am contemplating not linking and unlinking
but simply when walk whole lists each time.

Added (maybe temporary) version of DriveInfo that tries to walk through partitions of a drive and print out information.

*Extended USBDrive::findPartition*, including:
a) support extended partitions
b) reworked the GPT walking to not bail on first not known GUID but instead all GUIDS, until empty entry found
b1) The enumeration does not check the guid.  Instead returns it and lets caller decide if they understand the guid
c) returns a little more information needed for format.
d) return code was updated from bool to int, where 0 is no FS found, and other values to know if MBR, GPT, EXT partition

*Enumeration*:  Added method to USBDrive::startFilesystems

Which first does begin() to make sure drive is ready, it then keeps calling findPartition until it receives
back a return of 0.

It then calls off to claim for each of the free USBFilesystem objects until one claims it or none...
If FS claimed it removes the item from list and adds to the drives list.

When drive goes away, it walks it's list and calls another member of the new FS base class which the current
one calls end().  It also then removes the item from drive list and puts back on free.