Add Task() support for automatic FS Starting when drive inserted
Added support where when you insert an USB Drive in (or at sketch startup),
The using sketch does not have to do anything to get the drive to enumerate the filesystems,
it will happen by default when the next call to myUSB.Task() is called.

The user can set this to manual, and then there are is another method to call that will do this
for all drives that are defined.  Or you can still do it on a drive baises.

This first part made it reasonably automatic for those sketches that use one or more of the
USBDrive and USBFilesystem objects and just have it work.

In addition, I did some cleanup of the USBFSBase and USBFilesystem classes.
In particular I made the USBFSBase derive from FS, and as such then the
USBFilesystem class simply derive from this class.

I also changed the linked list of Filesystems, from moving the objects from a
free pool to a link list per driver, to simply keep a list of all FS, and when
one is claimed we know it as we update the device member anyway.  Downside is that when new drives connect or disconnect, we have to scan the full list.
Upside it gives us a list of Filesystems that we can walk.

Also we have simple state information stored in each USBFSBase object that the
state information has changed.  Which includes its connection state and if it
was formatted.  In addition there is a global (static var) that you can query
to see if any of these File Systems have changed state.

This code is used in the new MTP loop callback code, to now be able to detect
the that list of active filesystems has changed, and this code will then
add or remove the item to the MTP storage list.