Update USBHost_t36.h
I confirmed by building one of the storage Example sketches that the code to get/set File dates for files in USBDrive/Filesystem was not compiled into the programs and as such used the default implemention that did not get or set dates.

This was caused by while we were developing it it was under the
#iffdef FS_FILE_SUPPORT_DATES

Which we add into FS.h.

It was later removed from FS.h

So I changed this test to check for version of Teensyduino.

#if TEENSYDUINO >= 156