Part 1 - Allow some Keyboard Boot VID/PID to be claimed by USBHIDParser
I am running into a case with the Logitech G Pro Wireless mouse where if a sort of phantom keyboard part of the object was allowed to be claimed by Keyboard, it will kill the Mouse if nothing was processed for a few minutes.

The code now starts off with a VID/PID list part of Keyboard object, that both the Keyboard objects check for as well as USBHidParsers check when HID Boot Keyboard, and if in this list, it allows the HID parser to claim it instead of the Keyboard controller.

Part 2 - will have some class that can process the Keyboard HID data.