Updated and new examples for testing and using keyboards
The HIDDumper - code was updated to work better with showing keyboard data.

The keyboard_viewer - shows a psuedo keyboard on an ILI9341 display, and changes the color of which keys are pressed, helps visualize that the N key rollover is working.  Also shows last key pressed or released, including multi-media keys

The MouseKeyboardForward - is like KeyboardForward - but also shows debug info if not configured to usb type including keyboard (or mouse)
It is setup with two keyboards objects to help debug with Logitech Pro G wireless mouse which created a phantom keyboard.