Update Storage Examples
Several of the examples were not updated to work with the now automatic enumeration of partitions code.

I removed a few more pieces of the old manual code and updated the examples to reflect this.

I also updated the wave player example to play all of the wave files in the root directory of the storage instead of some fixed named ones.