USBHost::queue_Data_Transfer - always disable USB IRQ
Often times, the code is already run within an ISR and as such the
internal data structures will not be corrupted by another ISR.

But there were other cases were code is called from user code (not within this ISR), in which ase the transfer lists could be corrupted if we are in the middle of an update and ISR happens which also tries to change the list.

Code was updated to check to see if the USB irq was enabled on entry and then disables the IRQ. And when it returns if the IRQ was enabled on entry it should reenable it on return.