New USBSerial example sketch plus...
I made a version of the USBtoSerial like teensy example that instead does USB Seruak to USBHost Serial.

Put in optional constructors and the like in comments to allow users to have more knowledge of the options.

I also now added in support to the serial objects, to be able to call begin earlier on (before we actually detect the USB Host serial object and specify the Baud rate.   This was an issue in an earlier forum thread.

This sketch also looks at the current baud rate of the Serial object and tries to update the USBHost one.  Note: I don't update if I find Serial object baud is 0, which I am seeing in the Arduino 2.1.0 Serial monitor.

I also added in optional support the USB Serial constructors to allow you to pass in a VID, PID, what it maps to and if on Interface or whole object, to allow people an easier time to try seeing if they can easily add in their device which underlying it might be a known Serial converter chip but with different VID/PID.

Have done some testing but probably can use more