SerialEMU fixes - There was a buffer problem
There was a problem when we received a message from the plugged in device over seremu we screwed up the first or last character of the buffer.

Updated the code to use similar code to the USBSerial class code.

Also  added a begin() and end() method that does nothing, but makes it easier to swap in the emu class to the normal serial.

Also the code was still always outputting some debug messages.  Now they should all be under #ifdef control, using DBGPrintf macro/dummy function.