Remove erroneous comment in InputFunctions.ino
I'm pretty sure the comment specifying that the Teensy needs to be in USB MIDI mode is a copy-and-paste mistake from the Interface_16x16 example. That example hosts a bunch of devices and then forwards the MIDI commands to the Teensy's host, so the Teensy needs to present as a MIDI device to its host. This example just prints out MIDI info, though, so it should just be in standard Serial mode.